modify home markup and style
