+++ mclient/src/components/home/home.module.scss
@@ -1,0 +1,114 @@
+.container {
+    padding: 1.2rem;
+    margin-top: 7rem;
+    margin-bottom: 10rem;
+    z-index: 10;
+    left: 0px;
+    display: flex;
+    flex-direction: column;
+    background: linear-gradient(146.6deg, #1f2328 12.49%, #17191d 81%);
+    box-shadow: -8px -8px 20px rgba(42, 46, 53, 0.58886),
+        3px 8px 20px rgba(0, 0, 0, 0.353065);
+    border-radius: 37px 37px 0px 0px;
+
+    a:hover {
+        color: var(--title);
+    }
+
+    .menu {
+        display: none;
+    }
+
+    .recentlyPlayed p,
+    .trending p {
+        color: var(--subtitle);
+    }
+
+    .trending p {
+        margin-top: 1.2em;
+    }
+
+    .cardsGroup {
+        display: grid;
+        grid-template-columns: 1fr 1fr;
+        gap: 10px;
+
+        .card {
+            padding: 1em;
+            background-color: var(--color-dark-lighten);
+            box-shadow: -1px -5px 12px rgba(42, 46, 53, 0.58886);
+        }
+
+        img {
+            border-radius: 50%;
+        }
+
+        .album {
+            margin: 1em 0 0;
+            color: var(--title);
+            text-align: center;
+
+            p {
+                line-height: normal;
+            }
+        }
+    }
+
+    @media screen and (min-width: 768px) {
+        display: grid;
+        grid-template-columns: repeat(3, 1fr);
+        width: 73%;
+        margin-left: 25%;
+
+        .menu {
+            position: absolute;
+            left: 1.2em;
+            grid-column: 1 / 2;
+            color: var(--subtitle);
+            display: flex;
+            flex-direction: column;
+            gap: 20px;
+
+            li {
+                margin-bottom: 20px;
+            }
+
+            .divider {
+                position: relative;
+                width: 100%;
+                height: 3px;
+                left: 0;
+                top: 0;
+
+                background: var(--subtitle);
+                box-shadow: inset 0px 1px 2px rgba(0, 0, 0, 0.593231);
+                border-radius: 2px;
+            }
+        }
+
+        .cardsGroup {
+            gap: 2em;
+            grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
+        }
+
+        .recentlyPlayed,
+        .trending {
+            grid-column: 1 / -1;
+        }
+    }
+
+    @media screen and (min-width: 1216px) {
+        width: 80%;
+        margin-left: 14rem;
+    }
+}
+
+.playingNow {
+    position: fixed;
+    bottom: 0;
+    left: 0;
+    width: 100vw;
+    height: 10rem;
+    border-top: 3px solid var(--color-dark);
+    background-color: var(--color-dark);
+}
